Rename Grains.Impl to Grains and update references
