+++ src/Grains/Grains.csproj
@@ -14,7 +14,13 @@
 
     <ItemGroup>
         <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="8.1.2"/>
+        <PackageReference Include="Microsoft.Orleans.CodeGenerator" Version="8.2.0">
+          <PrivateAssets>all</PrivateAssets>
+          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+        <PackageReference Include="Microsoft.Orleans.Core.Abstractions" Version="8.2.0" />
         <PackageReference Include="Microsoft.Orleans.Runtime" Version="8.2.0"/>
+        <PackageReference Include="Microsoft.Orleans.Sdk" Version="8.2.0" />
         <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="8.1.2"/>
     </ItemGroup>
